requirements.txt
Flask==2.2.5
oauthlib==3.2.2
gunicorn==20.1.0
requests==2.31.0
pyOpenSSL==23.2.0
Flask-Login==0.6.2
Flask-Bootstrap==3.3.7
Werkzeug==2.3.0
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
flask_wtf==1.2.1
flask_sqlalchemy==3.1.1
flask_migrate==4.0.5
flask_bcrypt==1.0.1